fix FE tests
- extract fetchSuggestions into a separate module to break circular dependency
- handle headers not existing in API response (no headers in tests)
